+++ get_word_lists.py
@@ -1,18 +1,24 @@
 import random
 
-def main():
-    wordlists = create_wordlists()
+# def main():
+#     word_lists = create_word_lists()
 
-def create_wordlists():
-    wordlists = {i: [] for i in range(3, 16)}
-    with open("words_alpha.txt") as file:
-        for word in file:
-            word = word.strip()
-            if 3 <= len(word) <= 15:
-                wordlists[len(word)].append(word)
-    for length in wordlists:
-        random.shuffle(wordlists[length])
-    return wordlists
+class GetWordList:
 
-if __name__ == "__main__":
-    main()
+    def __init__(self, filename="words_alpha.txt"):
+        """
+        Initializes the WordListManager with the path to the word list file.
+        """
+        self.filename = filename
+        self.word_lists = self.create_word_lists()
+
+    def create_word_lists(self):
+        word_lists = {i: [] for i in range(3, 16)}
+        with open("words_alpha.txt") as file:
+            for word in file:
+                word = word.strip()
+                if 3 <= len(word) <= 15:
+                    word_lists[len(word)].append(word)
+        for length in word_lists:
+            random.shuffle(word_lists[length])
+        return word_lists
